Add "see more" link to GHE-not-supported warning (#609)
* Add "see more" link to GHE-not-supported warning

I lived for several months thinking that support for caching action on GHE is just a matter of time, because it's such an important thing to have. Only today, I discovered that originally it was not planned at all. And that people already created some workarounds. So I hope that linking the issue from the warning message will save other people from what happened to me :-)

* Update new GHE-not-supported message in tests

* Update generated dist files